Fixed spelling issues.
